--- src/components/layout.js
@@ -1,19 +1,0 @@
-import React from "react"
-import PropTypes from "prop-types"
-import Profile from "./Profile"
-
-const Layout = ({ children }) => {
-  return (
-    <>
-      <Profile />
-      <aside></aside>
-      <main>{children}</main>
-    </>
-  )
-}
-
-Layout.propTypes = {
-  children: PropTypes.node.isRequired,
-}
-
-export default Layout
